Putting Finite State Machine as Nested Class
This redesign puts PolyFsm as a nested class of InToPost, and InToPost as a nested class of ExprEval.

This is meant to have more encapsulation in the reading of the code.

Also, the methods in these three classes are no longer static, and contructors were created to account for this.